+++ tests/test_outline.py
@@ -156,12 +156,12 @@ def test_outline_section_not_found(cratedb_outline_builtin):
     assert section_not_found is None
 
 
-def test_outline_section_items_as_dict(cratedb_outline_builtin):
-    items = cratedb_outline_builtin.find_items(section_name="Docs", as_dict=True)
+def test_outline_section_items_dict(cratedb_outline_builtin):
+    items = cratedb_outline_builtin.find_items(section_name="Docs").to_dict()
     assert items[0]["title"] == "CrateDB README"
 
 
-def test_outline_section_items_as_objects(cratedb_outline_builtin):
+def test_outline_section_items_objects(cratedb_outline_builtin):
     items = cratedb_outline_builtin.find_items(section_name="Docs")
     assert items[0].title == "CrateDB README"
 
@@ -177,12 +177,12 @@ def test_outline_section_all_items(cratedb_outline_builtin):
     assert len(items) >= 30
 
 
-def test_outline_find_items_as_dict(cratedb_outline_builtin):
-    items = cratedb_outline_builtin.find_items(title="gtfs", as_dict=True)
+def test_outline_find_items_dict(cratedb_outline_builtin):
+    items = cratedb_outline_builtin.find_items(title="gtfs").to_dict()
     assert "Capture GTFS and GTFS-RT data" in items[0]["description"]
 
 
-def test_outline_find_items_as_objects(cratedb_outline_builtin):
+def test_outline_find_items_objects(cratedb_outline_builtin):
     items = cratedb_outline_builtin.find_items(title="gtfs")
     assert "Capture GTFS and GTFS-RT data" in items[0].description