+++ CHANGES.md
@@ -8,6 +8,8 @@
   and `--optional`, to directly convert/expand the source outline file into
   an `llms.txt` file.
 - llms-txt: Renamed subcommand `build` to `bundle --format=llms-txt`
+- Outline: Refactored quirky `as_dict` parameter to more fluent `.to_dict()`
+  interface
 
 ## v0.0.3 - 2025-05-10
 - Outline: Refactored the source of truth for the documentation outline